Set win lose but hero bug

+++ OpenDota/app/src/main/java/vn/edu/usth/opendota/models/Heroes.java
@@ -1,36 +1,36 @@
 package vn.edu.usth.opendota.models;
 
 public class Heroes {
-    private long hero_id;
-    private long last_played;
-    private long games;
-    private long win;
-    private long with_games;
-    private long with_win;
-    private long against_games;
-    private long against_win;
+    private int hero_id;
+    private int last_played;
+    private int games;
+    private int win;
+    private int with_games;
+    private int with_win;
+    private int against_games;
+    private int against_win;
 
-    public long getHeroID() { return hero_id; }
-    public void setHeroID(long value) { this.hero_id = hero_id; }
+    public int getHeroID() { return hero_id; }
+    public void setHeroID(int value) { this.hero_id = hero_id; }
 
-    public long getLastPlayed() { return last_played; }
-    public void setLastPlayed(long value) { this.last_played = last_played; }
+    public int getLastPlayed() { return last_played; }
+    public void setLastPlayed(int value) { this.last_played = last_played; }
 
-    public long getGames() { return games; }
-    public void setGames(long value) { this.games = value; }
+    public int getGames() { return games; }
+    public void setGames(int value) { this.games = value; }
 
-    public long getWin() { return win; }
-    public void setWin(long value) { this.win = value; }
+    public int getWin() { return win; }
+    public void setWin(int value) { this.win = value; }
 
-    public long getWithGames() { return with_games; }
-    public void setWithGames(long value) { this.with_games = value; }
+    public int getWithGames() { return with_games; }
+    public void setWithGames(int value) { this.with_games = value; }
 
-    public long getWithWin() { return with_win; }
-    public void setWithWin(long value) { this.with_win = value; }
+    public int getWithWin() { return with_win; }
+    public void setWithWin(int value) { this.with_win = value; }
 
-    public long getAgainstGames() { return against_games; }
-    public void setAgainstGames(long value) { this.against_games = against_games; }
+    public int getAgainstGames() { return against_games; }
+    public void setAgainstGames(int value) { this.against_games = against_games; }
 
-    public long getAgainstWin() { return against_win; }
-    public void setAgainstWin(long value) { this.against_win = against_win; }
+    public int getAgainstWin() { return against_win; }
+    public void setAgainstWin(int value) { this.against_win = against_win; }
 }
